Failures down to double digit numbers

+++ nimble/core/data/matrixAxis.py
@@ -117,8 +117,8 @@ class MatrixAxis(Axis, metaclass=ABCMeta):
         baseDtype = self._base._data.dtype
         if baseDtype != np.object_ and retDtype == np.object_:
             self._base._data = self._base._data.astype(np.object_)
-        elif baseDtype == np.int_ and retDtype == np.float:
-            self._base._data = self._base._data.astype(np.float)
+        elif baseDtype == np.int_ and retDtype == np.float_:
+            self._base._data = self._base._data.astype(np.float_)
         elif baseDtype == np.bool_ and retDtype != np.bool_:
             self._base._data = self._base._data.astype(retDtype)
